+++ backend/runtimes/node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cat-factory/node-server
 
+## 0.57.1
+
+### Patch Changes
+
+- Updated dependencies [6da6637]
+  - @cat-factory/server@0.64.3
+
 ## 0.57.0
 
 ### Minor Changes